tests/wpt/web-platform-tests/2dcontext/fill-and-stroke-styles/2d.gradient.interpolate.zerosize.stroke.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.gradient.interpolate.zerosize.stroke</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.gradient.interpolate.zerosize.stroke</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);

var g = ctx.createLinearGradient(50, 25, 50, 25); // zero-length line (undefined direction)
g.addColorStop(0, '#f00');
g.addColorStop(1, '#f00');
ctx.strokeStyle = g;
ctx.rect(20, 20, 60, 10);
ctx.stroke();
_assertPixel(canvas, 19,19, 0,255,0,255, "19,19", "0,255,0,255");
_assertPixel(canvas, 20,19, 0,255,0,255, "20,19", "0,255,0,255");
_assertPixel(canvas, 21,19, 0,255,0,255, "21,19", "0,255,0,255");
_assertPixel(canvas, 19,20, 0,255,0,255, "19,20", "0,255,0,255");
_assertPixel(canvas, 20,20, 0,255,0,255, "20,20", "0,255,0,255");
_assertPixel(canvas, 21,20, 0,255,0,255, "21,20", "0,255,0,255");
_assertPixel(canvas, 19,21, 0,255,0,255, "19,21", "0,255,0,255");
_assertPixel(canvas, 20,21, 0,255,0,255, "20,21", "0,255,0,255");
_assertPixel(canvas, 21,21, 0,255,0,255, "21,21", "0,255,0,255");


});
</script>
